Moving pagers to its own package.
