+++ src/test/java/com/github/jambodb/storage/pagers/FilePagerTests.java
@@ -1,7 +1,8 @@
-package com.github.jambodb.storage.btrees;
+package com.github.jambodb.storage.pagers;
 
-import com.github.jambodb.storage.btrees.files.FileBTreePage;
-import com.github.jambodb.storage.btrees.files.FilePager;
+import com.github.jambodb.storage.btrees.Serializer;
+import com.github.jambodb.storage.pagers.FileBTreePage;
+import com.github.jambodb.storage.pagers.FilePager;
 import com.github.jambodb.storage.btrees.mock.MockObject;
 import com.github.jambodb.storage.btrees.mock.MockObjectSerializer;
 import com.github.jambodb.storage.btrees.mock.StringSerializer;
